Salary search, Salary submission enhancement

+++ api/TechSalary/SalarySubmissionApi/Models/SalarySubmission.cs
@@ -7,7 +7,7 @@ namespace SalarySubmissionApi.Models
         public string Company { get; set; } = string.Empty;
         public string Role { get; set; } = string.Empty;
         public string ExperienceLevel { get; set; } = string.Empty;
-        public int SalaryAmount { get; set; }
+        public long SalaryAmount { get; set; }
         public string Currency { get; set; } = string.Empty;
         public bool Anonymize { get; set; }
         public string Status { get; set; } = "PENDING";
